+++ tests/testthat/test-allelic_series_sumstats.R
@@ -35,7 +35,6 @@ test_that("Check COAST on sumstats.", {
 
 # ------------------------------------------------------------------------------
 
-
 test_that("Check COAST with missing inputs.", {
   
   # Null setting.
@@ -108,7 +107,7 @@ test_that("Check genomic inflation factor.", {
 
 test_that("Check case of rank-deficient LD.", {
   
-  withr::local_seed(101)
+  withr::local_seed(102)
   data <- DGP(n = 1e2, prop_causal = 0)
   sumstats <- CalcSumstats(data = data)
   ld <- sumstats$ld
@@ -169,40 +168,142 @@ test_that("Check cases where not all annotaiton categories are provided.", {
   }
   
   RunCOAST <- function(sumstats) {
-    result <- COASTSS(
+    result <- suppressWarnings(
+      COASTSS(
+        anno = sumstats$anno,
+        beta = sumstats$sumstats$beta,
+        se = sumstats$sumstats$se,
+        ld = sumstats$ld,
+        maf = sumstats$maf
+      )
+    )
+    pvals <- result@Pvals
+    return(pvals)
+  }
+  
+  # Exclude category 1.
+  data0 <- ExcludeAnno(data, anno = 1)
+  sumstats0 <- CalcSumstats(data = data0)
+  result0 <- RunCOAST(sumstats0)
+  expect_true(all(result0$pval >= 0.05))
+  
+  # Exclude category 2.
+  data1 <- ExcludeAnno(data, anno = 2)
+  sumstats1 <- CalcSumstats(data = data1)
+  result1 <- RunCOAST(sumstats1)
+  expect_true(all(result1$pval >= 0.05))
+  
+  # Exclude category 3.
+  data2 <- ExcludeAnno(data, anno = 3)
+  sumstats2 <- CalcSumstats(data = data2)
+  result2 <- RunCOAST(sumstats2)
+  expect_true(all(result2$pval >= 0.05))
+  
+  # Exclude two categories.
+  data12 <- ExcludeAnno(data1, anno = 3)
+  expect_true(all(data12$anno == 1))
+  sumstats12 <- CalcSumstats(data = data12)
+  result12 <- RunCOAST(sumstats12)
+  expect_true(all(result12$pval >= 0.05))
+  
+})
+
+# ------------------------------------------------------------------------------
+
+test_that("Check COASTSS runs with zero-based annotations.", {
+  
+  withr::local_seed(103)
+  
+  WrapCOASTSS <- function(data, weights) {
+    sumstats <- CalcSumstats(data = data)
+    test <- COASTSS(
       anno = sumstats$anno,
       beta = sumstats$sumstats$beta,
       se = sumstats$sumstats$se,
       ld = sumstats$ld,
       maf = sumstats$maf
     )
-    pvals <- result@Pvals
-    return(pvals)
+    pvals <- test@Pvals
+    p_omni <- pvals$pval[pvals$test == "omni"]
+    return(p_omni)
   }
   
-  # Exclude category 0.
-  data0 <- ExcludeAnno(data, anno = 0)
-  sumstats0 <- CalcSumstats(data = data0)
-  result0 <- RunCOAST(sumstats0)
-  expect_true(all(result0$pval >= 0.05))
+  # Null. 
+  data <- DGP(prop_causal = 0)
+  data$anno <- data$anno - 1
+  suppressWarnings(
+    expect_warning(p <- WrapCOASTSS(data))
+  )
+  expect_gt(p, 0.05)
   
-  # Exclude category 1.
-  data1 <- ExcludeAnno(data, anno = 1)
-  sumstats1 <- CalcSumstats(data = data1)
-  result1 <- RunCOAST(sumstats1)
-  expect_true(all(result1$pval >= 0.05))
-  
-  # Exclude category 2.
-  data2 <- ExcludeAnno(data, anno = 2)
-  sumstats2 <- CalcSumstats(data = data2)
-  result2 <- RunCOAST(sumstats2)
-  expect_true(all(result2$pval >= 0.05))
-  
-  # Exclude two categories.
-  data12 <- ExcludeAnno(data1, anno = 2)
-  expect_true(all(data12$anno == 0))
-  sumstats12 <- CalcSumstats(data = data12)
-  result12 <- RunCOAST(sumstats12)
-  expect_true(all(result12$pval >= 0.05))
+  # Alternative. 
+  data <- DGP(prop_causal = 1)
+  data$anno <- data$anno - 1
+  suppressWarnings(
+    expect_warning(p <- WrapCOASTSS(data))
+  )
+  expect_lt(p, 0.05)
   
 })
+
+# ------------------------------------------------------------------------------
+
+test_that("Check COASTSS with different numbers of categories.", {
+  
+  WrapCOASTSS <- function(data, weights) {
+    sumstats <- CalcSumstats(data = data)
+    test <- suppressWarnings({COASTSS(
+      anno = sumstats$anno,
+      beta = sumstats$sumstats$beta,
+      se = sumstats$sumstats$se,
+      ld = sumstats$ld,
+      maf = sumstats$maf,
+      weights = weights
+    )})
+    pvals <- test@Pvals
+    p_omni <- pvals$pval[pvals$test == "omni"]
+    return(p_omni)
+  }
+  
+  withr::local_seed(101)
+  
+  # 2 categories.
+  ## Null.
+  data <- DGP(
+    beta = c(0, 0),
+    prop_anno = c(1, 1),
+    weights = c(1, 1)
+  )
+  p <- WrapCOASTSS(data, weights = c(1, 2))
+  expect_gt(p, 0.05)
+  
+  ## Alternative.
+  data <- DGP(
+    beta = c(1, 2),
+    prop_anno = c(1, 1),
+    weights = c(1, 1)
+  )
+  p <- WrapCOASTSS(data, weights = c(1, 2))
+  expect_lt(p, 0.05)
+  
+  # 4 categories.
+  ## Null.
+  data <- DGP(
+    beta = c(0, 0, 0, 0),
+    prop_anno = c(1, 1, 1, 1),
+    weights = c(1, 1, 1, 1)
+  )
+  p <- WrapCOASTSS(data, weights = c(1, 2, 3, 4))
+  expect_gt(p, 0.05)
+  
+  ## Alternative.
+  data <- DGP(
+    beta = c(1, 2, 3, 4),
+    prop_anno = c(1, 1, 1, 1),
+    weights = c(1, 1, 1, 1)
+  )
+  p <- WrapCOASTSS(data, weights = c(1, 2, 3, 4))
+  expect_lt(p, 0.05)
+  
+})
+